Reorganize package layout
