+++ src/fs/node-package-fs.ts
@@ -3,11 +3,11 @@ import { createRequire } from 'node:module';
 import path from 'node:path';
 import process from 'node:process';
 import { fileURLToPath, pathToFileURL } from 'node:url';
-import { PackageInfo } from '../package-info.js';
-import { PackageJson, isValidPackageJson } from '../package.json.js';
-import { PackageFS } from './package-fs.js';
+import { PackageInfo } from '../package/package-info.js';
+import { PackageJson, isValidPackageJson } from '../package/package.json.js';
 import { Import } from '../resolution/import.js';
 import { PackageResolution } from '../resolution/package-resolution.js';
+import { PackageFS } from './package-fs.js';
 
 /**
  * Node.js-specific implementation of package file system.
