+++ src/mod.ts
@@ -2,8 +2,5 @@
  * @module @run-z/ndk
  */
 export * from './fs/mod.js';
-export * from './package-entry-point.js';
-export * from './package-entry-targets.js';
-export * from './package-info.js';
-export * from './package.json.js';
+export * from './package/mod.js';
 export * from './resolution/mod.js';
