+++ src/package/package-info.ts
